+++ Produtos.cs
@@ -3,14 +3,14 @@ namespace LinqPratice;
 
 abstract class Produtos
 {
-    public Produtos(string nome, decimal preco)
+    public Produtos(string nome, int preco)
     {
         Nome = nome;
         Preco = preco;
     }
 
     public string Nome { get; }
-    public decimal Preco { get; }
+    public int Preco { get; }
 
 
     public virtual void ExibirDados()
